restructure matrix designer
